Add files via upload
This code allows you to create a graphical application using the Tkinter library to build the user interface and OpenCV to work with images. This application aims to provide a set of tools and features that enable the user to make modifications and edits to images, such as uploading images, adding effects, and drawing shapes on the image. PIL is used to transform images and load them into the user interface correctly.